pyproject.toml
[project]
name = "smart-receipt-assistant"
version = "0.1.0"
description = "智能票据报销助手 - PaddleOCR + LangChain + ERINE"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "openai>=1.0.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.1",
    "langchain-openai>=0.1.0",
    "requests>=2.31.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "rich>=13.0.0",
    "pillow>=10.0.0",
    "gradio>=4.0.0",
    "openpyxl>=3.1.0",
    "paddleocr>=3.4.0",
    "paddlepaddle>=3.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "jupyter>=1.0.0",
]

[project.scripts]
receipt-assistant = "src.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]
